IS-3648: CommonJS to ESM
