+++ package.json
@@ -3,12 +3,13 @@
   "version": "0.2.0",
   "private": true,
   "description": "Digitalisering av sykefraværsoppfølging",
+  "type": "module",
   "main": "index.tsx",
   "test": "pnpm test",
   "start": "pnpm start",
   "scripts": {
-    "start": "webpack serve --config webpack.dev.ts",
-    "build": "webpack --config webpack.prod.ts && pnpm run build:server",
+    "start": "webpack serve --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js && pnpm run build:server",
     "build:server": "tsc --build ./server/tsconfig.json",
     "lint-and-typecheck": "pnpm lint && pnpm typecheck",
     "typecheck": "tsc --noEmit -p tsconfig.json",
@@ -47,7 +48,7 @@
     "classnames": "2.5.1",
     "core-js": "3.44.0",
     "dayjs": "1.11.13",
-    "dotenv": "8.6.0",
+    "dotenv": "16.5.0",
     "dotenv-expand": "5.1.0",
     "express": "4.22.1",
     "express-http-proxy": "2.1.2",
