+++ server/tsconfig.json
@@ -1,10 +1,14 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
     "rootDir": "..",
     "outDir": "../dist-server",
     "composite": true,
-    "tsBuildInfoFile": "../dist-server/.tsbuildinfo"
+    "tsBuildInfoFile": "../dist-server/.tsbuildinfo",
+    "noEmit": false,
+    "allowImportingTsExtensions": false
   },
   "include": ["../server.ts", "./**/*.ts"],
   "exclude": [
