+++ webpack.common.js
@@ -1,13 +1,16 @@
 import path from 'path';
-import { Configuration } from 'webpack';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import { CleanWebpackPlugin } from 'clean-webpack-plugin';
 import Dotenv from 'dotenv-webpack';
 import TsconfigPathsPlugin from 'tsconfig-paths-webpack-plugin';
+import { fileURLToPath } from 'url';
 
 const extensions = ['.tsx', '.jsx', '.js', '.ts', '.json'];
 
-const commonConfig: Configuration = {
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const commonConfig = {
   entry: path.resolve(__dirname, 'src', 'index.tsx'),
   output: {
     filename: '[name].bundle.js',
